+++ packages/eth2.0-utils/src/converters/index.ts
@@ -1,1 +1,2 @@
 export * from "./json";
+export * from "./yaml";
